Move `dumpCcSamples` to its owm module
